refactor: common draggable list component
