+++ frontend/src/components/DraggableList.vue
